refactor(mcp): one module per /mcp tool so Phase 1 tool work lands in separate files
Mechanical split of plugin/host/mcp/tools.js ahead of the Phase 1 fan-out
(ae_exec semantics / ae_status+previewFrame / ae_read run in parallel and would
all collide on one registry file):

- mcp/tools/status.js, mcp/tools/exec.js: each exports { definition, call(args, context, deps) }
- mcp/tool-result.js: textResult / noTopLevelCombinator shared without a registry cycle
- mcp/tools.js: aggregates TOOL_MODULES, rejects top-level combinators and duplicate
  names, dispatches by name; list order and the public buildTools/list/call API are unchanged
- cep-runtime-contract.test.js: the new files join CEP_EXECUTED_FILES (node:-specifier and
  require-graph guards now cover them) and the local-require scanner also follows ../ paths
  so subdirectory tool modules cannot slip past it

No behavior change. host: 126 tests, 109 pass, 1 fail (pre-existing local-only symlink
EPERM), 16 skipped.
